+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ unreleased
 
 * Remove the PasteDeploy FakeApp package in favor of ``montague_testapps``.
 * Enable a ``looponfail`` tox environment.
+* Reincorporate the test JSON config loader, active only during tests.
 
 0.1.5 (2015-05-12)
 -----------------------------------------